Refactoring docstring parsing modules
