+++ src/safeds_stubgen/docstring_parsing/_abstract_docstring_parser.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from mypy import nodes
 
-    from safeds_stubgen.api_analyzer import Class, ParameterAssignment
+    from safeds_stubgen.api_analyzer import Class
 
     from ._docstring import (
         AttributeDocstring,
