+++ src/safeds_stubgen/docstring_parsing/_create_docstring_parser.py
@@ -2,12 +2,12 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from griffe.enumerations import Parser
+
+from ._docstring_parser import DocstringParser
 from ._docstring_style import DocstringStyle
 from ._epydoc_parser import EpydocParser
-from ._googledoc_parser import GoogleDocParser
-from ._numpydoc_parser import NumpyDocParser
 from ._plaintext_docstring_parser import PlaintextDocstringParser
-from ._restdoc_parser import RestDocParser
 
 if TYPE_CHECKING:
     from ._abstract_docstring_parser import AbstractDocstringParser
@@ -17,10 +17,10 @@ def create_docstring_parser(style: DocstringStyle) -> AbstractDocstringParser:
     if style == DocstringStyle.EPYDOC:
         return EpydocParser()
     elif style == DocstringStyle.GOOGLE:
-        return GoogleDocParser()
+        return DocstringParser(Parser.google)
     elif style == DocstringStyle.NUMPYDOC:
-        return NumpyDocParser()
+        return DocstringParser(Parser.numpy)
     elif style == DocstringStyle.REST:
-        return RestDocParser()
+        return DocstringParser(Parser.sphinx)
     else:
         return PlaintextDocstringParser()
